Cambio en precio final  a tipo "float".

+++ Final1096917/Final1096917/Electrodomestico.h
@@ -15,7 +15,7 @@ public:
 	void setConsumo(string eConsumo);
 	int getPeso();
 	void setPeso(int ePeso);
-	virtual int calcularPrecio();
+	virtual float calcularPrecio();
 	~Electrodomestico();
 private:
 	float precioBase;
